+++ CHANGELOG.md
@@ -47,6 +47,8 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - Biological Association UUID sent to GBIF was bad data (not a UUID)
 - PDF viewer is not rendering pages [#5023]
 - Can't unify two Taxon Names when they have the same relationship type from/to a same third Taxon Name [#4971]
+- A "successful" unify between two Taxon Names, one of which has a synonym, can leave the synonym in an unusable state [#4971]
+- Unify of two Asserted Distributions with the "same" Citation can cause loss of annotation data from the citation that gets destroyed [#4971]
 
 [#4937]: https://github.com/SpeciesFileGroup/taxonworks/issues/4937
 [#4944]: https://github.com/SpeciesFileGroup/taxonworks/issues/4944